strawberry/strawberry_1/backend/requirements.txt
Flask
Flask-CORS
Pillow
numpy
scikit-learn